apps/api/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
python-multipart>=0.0.6
httpx>=0.26.0,<1.0.0
google-genai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9
openai>=1.0.0

# Local ML
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Local OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Local Speech-to-Text
faster-whisper>=0.10.0

# Testing
pytest>=8.0.0
